code/mgo/mgo/requirements.txt
pyramid==1.5.1
PasteDeploy==1.5.3
translationstring==1.1
venusian==1.0a8
zope.deprecation==4.1.1
zope.interface==4.1.1
repoze.lru==0.6
WebOb==1.4
mgo==0.0
docutils==0.11
waitress==0.8.9
zope.sqlalchemy==0.7.4
transaction==1.4.3
SQLAlchemy==0.9.4
pyramid-tm==0.7
pyramid-debugtoolbar==2.1
pyramid-jinja2==2.3.2
Mako==1.0.0
Pygments==1.5
pyramid-mako==1.0.2
MarkupSafe==0.23
Jinja2==2.7.3
alembic==0.6.5
